pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pytorch_laplace"
version = "0.0.0"
description = "Pytorch-laplace package"
requires-python = ">=3.7.3"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python",
]
dependencies = [
    "black[jupyter]==23.3.0",
    "ruff==0.0.267",
    "typeguard==2.13.3",
    "pytest==7.1.2",
    "pytest-xdist==2.5.0",
    "pyright",
    "sphinx",
    "sphinx_rtd_theme",
    "sphinxcontrib-video",
    "torch>=1.13.1",
    "nnj@git+https://github.com/IlMioFrizzantinoAmabile/nnj",
    "backpack-for-pytorch",
]

[tool.setuptools.packages.find]
include = ["pytorch_laplace*"]

[tool.black]
# https://github.com/psf/black
line-length = 110
target-version = ["py38"]
exclude = "(.eggs|.git|.hg|.mypy_cache|.nox|.tox|.venv|.svn|_build|buck-out|build|dist|docs)"

[tool.isort]
profile = "black"
line_length = 110
force_sort_within_sections = "False"
order_by_type = "False"